requirements.txt
click==7.1.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.12.0
Werkzeug==1.0.0

pytest~=5.4.3
python-dateutil~=2.8.1
requests~=2.24.0